Removes censorship and tracking from the UI.
Also swaps out some of the RPC servers used since the Infura ones are known to censor.  Keydonix is just currently just CloudFlare's JSON-RPC server under the hood (DNS CNAME).  The reason the public CF server wasn't included was because there isn't a public testnet server, and I would like to ensure CF gets paid for the traffic I am sending it so they continue to offer their service (using the public gateway would result in the cost being entirely theirs).  Also, if CloudFlare censors (or ends up being too expensive), I can redirect Keydonix to QuickNode or somewhere else that is cheaper/censorship resistant.

The service worker tests were removed because the only thing that used service workers in production was the tracking code, so once that was ripped out the tests no longer passed (correctly).

The wallet test was disabled because after a day of debugging I was unable to figure out why changing the RPC server for mainnet from Infura to Keydonix results in the test failing.  When I tested the behavior in a production environment everything worked, so I decided to disable the test for now so I could move forward.

The linter on this project is incredibly strict, so while this PR is almost entirely just deletions, in many places I had to add/remove whitespace (newlines, indentation, etc.) to satisfy the linter.  Recommend viewing this changeset with whitespace changes ignored.

There is certainly more that could be done to further improve the privacy protection of Uniswap users, but I intentionally kept the scope here narrow.